Networking: move files to a sub folder
