+++ src/common/networking/networking_windows.c
@@ -3,7 +3,7 @@
 
 //Must be included after <mswsock.h>
 #include "fastfetch.h"
-#include "common/networking.h"
+#include "common/networking/networking.h"
 #include "util/stringUtils.h"
 #include "util/debug.h"
 
@@ -225,9 +225,8 @@ const char* ffNetworkingRecvHttpResponse(FFNetworkingState* state, FFstrbuf* buf
     char* headerEnd = NULL;
 
     do {
-        size_t availSpace = ffStrbufGetFree(buffer);
-        FF_DEBUG("Data reception loop #%d, current buffer size: %u, available space: %zu",
-                 ++recvCount, buffer->length, availSpace);
+        FF_DEBUG("Data reception loop #%d, current buffer size: %u, available space: %u",
+                 ++recvCount, buffer->length, ffStrbufGetFree(buffer));
 
         ssize_t received = recv(state->sockfd, buffer->chars + buffer->length, (int)ffStrbufGetFree(buffer), 0);
