+++ src/detection/weather/weather.c
@@ -1,4 +1,5 @@
 #include "weather.h"
+#include "common/networking/networking.h"
 
 #define FF_UNITIALIZED ((const char*)(uintptr_t) -1)
 static FFNetworkingState state;
