+++ src/detection/weather/weather.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common/networking.h"
+#include "fastfetch.h"
 
 void ffPrepareWeather(FFWeatherOptions* options);
 const char* ffDetectWeather(FFWeatherOptions* options, FFstrbuf* result);
